py_backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
httpx
pydantic
python-dotenv
